deleted old view activities and implemented fragment-based view switching in ReportCollectionActivity
