+++ app/src/main/java/mil/nga/dice/gridview/ReportGridFragment.java
@@ -7,7 +7,6 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
 import android.support.v4.content.LocalBroadcastManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -40,7 +39,7 @@ public class ReportGridFragment extends Fragment implements AdapterView.OnItemCl
 	
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-		final View v = inflater.inflate(R.layout.activity_report_grid, container, false);
+		final View v = inflater.inflate(R.layout.fragment_report_grid, container, false);
 		final GridView gridView = (GridView) v.findViewById(R.id.report_grid);
 		gridView.setAdapter(mReportsAdapter);
 		gridView.setOnItemClickListener(this);
