+++ app/src/main/res/menu/menu_report_collection.xml
@@ -2,15 +2,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:context="mil.nga.dice.ReportCollectionActivity">
 
-    <item android:id="@+id/list_button"
+    <item android:id="@+id/collection_view_list"
         android:icon="@drawable/ic_action_view_as_list"
         android:title="@string/show_map"
         android:showAsAction="ifRoom"/>
-    <item android:id="@+id/tiles_button"
+    <item android:id="@+id/collection_view_grid"
         android:icon="@drawable/ic_action_view_as_grid"
         android:title="@string/show_tiles"
         android:showAsAction="ifRoom"/>
-    <item android:id="@+id/map_button"
+    <item android:id="@+id/collection_view_map"
         android:icon="@drawable/ic_action_map"
         android:title="@string/show_map"
         android:showAsAction="ifRoom"/>
